Escape Everything (not just Strings).
Previously, the EscapeFilter would ignore anything that wasn't a String, this
meant that an object being printed could contain some xss content in its
toString() method (e.g standard java collections that contain a string with
xss) that didn't get escaped, even if the escape filter was used explicitly.
The EscapeFilter now converts anything to a String with toString(), and then
proceeds as normal.