Add PebbleEngine.Builder.build variant with custom base extensions
This fixes issue mbosecke/pebble#151, where default base extensions
(i.e. `i18n`) can be substituted for user-provided ones, in case there
are issues with the default set.

A new variant of `PebbleEngine.Builder.build` simply accepts a `List` of
`Extensions`, which will be used along with any `userProvidedExtensions`
otherwise given to `Builder`.

The existing API is unchanged, and returns the result of the new method,
but with the existing list of default extensions.