Allow the set tag to update existing values in parent scopes
This fixes a problem when using the set tag in a nested for tag. The for
tag pushes a new scope if there is already a "loop" variable in scope. If
a set tag executes in that nested scope, the changes are undone at the end
of the for tag, which is not what users expect.

Example:

{% set x = 1 %}
{% for i in 1..1 %}
  {% for j in 1..1 %}
    {% set x = 2 %}
  {% endfor %}
{% endfor %}
{{ x }}

Previously this would yield "1"; now it will yield "2".