Optimize ScopeChain for basic case with single scope
Motivation:

`ScopeChain#get` performs `containsKey` then `get` so it can account for null values.

The current implementation tries to optimize for the single scope case for the chain scanning with saving an Iterator.
It's possible to go further this way.

Modification:

Optimize for the most common use case where the key is defined and there's a single scope.
Only perform a `get` and only check `contains` if the value is null and there is more than one scope.

Result:

Better performance.