Optimize ResolvedAttribute
Motivation:

`ResolvedAttribute` is being introduced in 3.0 and is currently a SAM (could as well be a `Supplier<Object>`, all the more as they are usually implemented as Java lambdas.
The thing is all those lambdas come with a performance penalty, and there's actually not need for functions here. Functions would only be useful if evaluation was lazy, which is not the case.
What is needed here is only a wrapper around the resolved value (we need the wrapper because `AttributeResolver#resolve` can return null, which is different from a null resolved value.

Modification:

Turn ResolvedAttribute into a basic wrapper.

Result:

Better performance