Add native hints
In order to let any Java project including PebbleTemplates as a
dependency to be able to build and run a native image using the
AOT-compiler GraalVM, it is necessary to add native hints to the GraalVM
so that it can maintain class and variables information during
compile time (and consequently runtime) to be used as information for
Reflection, Proxies and other JVM mechanisms.

This will add the necessary TypeHints for the SpringBoot starter only,
in order to enable the building of a project using Spring Boot as
framework, and to promote further experiments with other frameworks.